iOS-1: Вынес сабвью в отдельные файлы
