+++ CakesHub/CakesHub/Streams/ProductStream/ProductDetail/ProductDetailScreen/Model/ProductDetailCells.swift
@@ -1,0 +1,13 @@
+//
+//  ProductDetailCells.swift
+//  CakesHub
+//
+//  Created by Dmitriy Permyakov on 19.03.2024.
+//
+
+import Foundation
+
+enum ProductDetailCells: String {
+    case ratingReviews = "Rating&Reviews"
+    case sellerInfo = "Seller info"
+}
