+++ CakesHub/CakesHub/Streams/ProductStream/ProductDetail/ProductDetailScreen/ViewModel/NewProductDetailModel+Mock.swift
